+++ internal/rad/rad.go
