+++ internal/tui/delete.go
